requirements.txt
flask==3.1.3
flask-cors==5.0.1
flask-jwt-extended==4.7.1
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
bcrypt==4.3.0
requests==2.32.3
python-dotenv==1.0.1
